Add onBatchStarted callback to "BatchCompleteListener".
This change adds a way for native modules to receive "batchStarted" event (I've actually renamed OnBatchCompleteListener to JSBatchListener so that the name does no longer suggest it can be used only for getting "batchComplete" events).

I think this callback will be important for implementing features that may want to interact with native modules from inside of the native code and will allow to isolate modifications of UI that are done by JS from the ones that can be driven by purely native code. A good example would be driving animations from native code. In that case we'd like for the animation code to be able to update view properties (such as transforms) but also props that affect shadow view hierarchy (layout-related props). Since changes made to layout props have to go though the regular prop update pipeline (update shadow view -> calculate layout on batch complete -> update values in UI thread) the best way IMO is to "trigger" onBatchComplete when animation frame is requested and let UIManager start all the view updating logic. The issue is with other updates that might be coming from JS at the same time. SO the solution is to detect if JS is in a process of sending native commands and wait until the process is finished - that's where this new callback I'm adding will help a lot.

Test plan:
Add boolean "batchState" variable in UImanagerModule. Update it to tru in #onBatchStarted and to false in #onBatchComplete. Add asserts in #onBatchComplete for "batchState" to be true when entering the method -> play with some apps, test reload, asserts should always be passing