Don't set JS responder lock on topScroll event.
This change unifies behaviour of dispatching CANCEL events across iOS and Android. The previous setup was for the scroll events to be special-cased in renderer configuration such that when topScroll got delivered the JS responder was set. This wasn't ideal because while it is easy for the scrollview to grab the responder it has no controll over releasing it. Therefore we were relying on the fact that while the user was scrolling we were also sending topTouchMove and topTouchEnd events. Also the order of delivering the events was crutial as the scrollview was expected to deliver last topScroll event before topTouchEnd.